Use pytest plugin for marker filtering
