+++ tests/conftest.py
@@ -1,10 +1,6 @@
 import os
 
 import pytest
-from _pytest.config import PytestPluginManager
-from _pytest.config.argparsing import Parser
-
-from localstack.testing.pytest.filters import filter_tests
 
 os.environ["LOCALSTACK_INTERNAL_TEST_RUN"] = "1"
 
@@ -51,69 +47,6 @@ def pytest_sessionstart(session):
         # this will be executed in the CLI tests as well, where we don't have the pytest_httpserver dependency
         # skip in that case
         pass
-
-
-@pytest.hookimpl
-def pytest_addoption(parser: Parser, pluginmanager: PytestPluginManager):
-    parser.addoption(
-        "--offline",
-        action="store_true",
-        default=False,
-        help="test run will not have an internet connection",
-    )
-
-
-def pytest_configure(config):
-    config.addinivalue_line(
-        "markers",
-        "skip_offline: mark the test to be skipped when the tests are run offline "
-        "(this test explicitly / semantically needs an internet connection)",
-    )
-    config.addinivalue_line(
-        "markers",
-        "only_on_amd64: mark the test as running only in an amd64 (i.e., x86_64) environment",
-    )
-    config.addinivalue_line(
-        "markers",
-        "only_on_arm64: mark the test as running only in an arm64 environment",
-    )
-    config.addinivalue_line(
-        "markers",
-        "only_in_docker: mark the test as running only in Docker (e.g., requires installation of system packages)",
-    )
-    config.addinivalue_line(
-        "markers",
-        "resource_heavy: mark the test as resource-heavy, e.g., downloading very large external dependencies, "
-        "or requiring high amount of RAM/CPU (can be systematically sampled/optimized in the future)",
-    )
-    config.addinivalue_line(
-        "markers",
-        "aws_validated: mark the test as validated / verified against real AWS",
-    )
-    config.addinivalue_line(
-        "markers",
-        "aws_only_localstack: mark the test as inherently incompatible with AWS, e.g. when testing localstack-specific features",
-    )
-    config.addinivalue_line(
-        "markers",
-        "aws_needs_fixing: test fails against AWS but it shouldn't. Might need refactoring, additional permissions, etc.",
-    )
-    config.addinivalue_line(
-        "markers",
-        "aws_manual_setup_required: validated against real AWS but needs additional setup or account configuration (e.g. increased service quotas)",
-    )
-    config.addinivalue_line(
-        "markers",
-        "aws_unknown: it's unknown if the test works (reliably) against AWS or not",
-    )
-    config.addinivalue_line(
-        "markers",
-        "multiruntime: parametrize test against multiple Lambda runtimes",
-    )
-
-
-def pytest_collection_modifyitems(config, items):
-    filter_tests(config, items)
 
 
 @pytest.fixture(scope="session")
